-Made changes in booking a trip
