+++ app/Http/Controllers/Api/BookingsController.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Api\BaseController;
 use App\Models\CustomerTrip;
 use App\Models\Organizer;
+use App\Models\Passenger;
 use App\Models\Trip;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class BookingsController extends BaseController
 {
@@ -84,4 +86,52 @@ class BookingsController extends BaseController
 
         return $this->sendResponse($booking,'This booking canceled successfully');
     }
+
+    public function bookTrip(Request $request)
+    {
+        error_log('Book trip request!');
+        $id = Auth::id();
+        $user = User::find($id);
+        if ($user == null) {
+            error_log('User not found!');
+            return $this->sendError('User not found!');
+        }
+
+        $validator = Validator::make($request->all(), [
+            'trip_id' => 'required',
+        ]);
+        if ($validator->fails()) {
+            error_log($validator->errors());
+            return $this->sendError('Validator failed! check the data', $validator->errors());
+        }
+        $trip = Trip::find($request['trip_id'])->with(['organizer' => function ($query) use ($id) {
+            $query->where('user_id', $id);
+        }])->first();
+        if ($trip->organizer != null) {
+            error_log('User is the one that created the trip!');
+            return $this->sendError('User is the one that created the trip!', [], 405);
+        }
+
+        $customerTrip = new CustomerTrip();
+        $customerTrip->trip()->associate($request['trip_id']);
+        $customerTrip->user()->associate($id);
+        $customerTrip->save();
+        $customerPassenger = new Passenger();
+        $customerPassenger->passenger_name= $user->first_name;
+        $customerPassenger->customerTrip()->associate($customerTrip->id);
+        $customerPassenger->save();
+        if($request['passengers']!=null) {
+            $passengers = $request['passengers'];
+
+            for ($i = 0; $i < count($passengers); $i++) {
+                $passenger = new Passenger;
+                $passenger->passenger_name = $passengers[$i]['name'];
+                $passenger->customerTrip()->associate($customerTrip->id);
+                $passenger->save();
+            }
+        }
+        error_log('book trip request succeeded!');
+        return $this->sendResponse($customerTrip, 'Succeeded!');
+    }
+
 }
